+++ addons/hw_posbox_homepage/controllers/__init__.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
+from . import jinja
 from . import main
+from . import technical
