+++ addons/hw_posbox_homepage/views/configure_wizard.jinja2
@@ -1,5 +1,5 @@
-{% extends "layout.html" %}
-{% from "loading.html" import loading_block_ui %}
+{% extends "_layout.jinja2" %}
+{% from "loading.jinja2" import loading_block_ui %}
 {% block head %}
     <script>
         $(document).ready(function () {
