+++ addons/hw_posbox_homepage/views/technical/credential.jinja2
@@ -1,5 +1,5 @@
-{% extends "layout.html" %}
-{% from "loading.html" import loading_block_ui %}
+{% extends "technical/_layout.jinja2" %}
+{% from "loading.jinja2" import loading_block_ui %}
 {% block head %}
     <script>
         var _onQueryDone = function () {
