+++ addons/hw_posbox_homepage/views/technical/general.jinja2
@@ -1,0 +1,5 @@
+{% extends "technical/_layout.jinja2" %}
+{% block content %}
+    <h2 class="text-center text-green">Technical Information</h2>
+    {{ iot_table(infos_table) }}
+{% endblock %}
